requirements.txt
Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
PyYAML==3.12
Werkzeug==0.14.1
argparse==1.2.1
click==6.7
ddt==1.1.1
fakeredis==0.9.0
funcsigs==1.0.2
google-api-python-client==1.6.5
httplib2==0.10.3
itsdangerous==0.24
mock==2.0.0
oauth2client==4.1.2
pbr==3.1.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
redis==2.10.6
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
wsgiref==0.1.2
